Extracted ScheduleRow class
