+++ src/guis/krow/guis/schedule_module/ScheduleRow.java
@@ -1,0 +1,136 @@
+package krow.guis.schedule_module;
+
+import java.io.IOException;
+
+import javafx.scene.control.Label;
+import javafx.scene.control.TableRow;
+import javafx.scene.input.MouseButton;
+import javafx.scene.layout.Background;
+import javafx.scene.layout.BackgroundFill;
+import javafx.scene.layout.VBox;
+import javafx.scene.paint.Color;
+import krow.guis.PopupHelper;
+import krow.guis.PopupHelper.PopupWrapper;
+import kröw.core.managers.WindowManager;
+import kröw.core.managers.WindowManager.NotSwitchableException;
+import zeale.guis.schedule_module.NewEvent;
+import zeale.guis.schedule_module.ScheduleModule;
+
+public class ScheduleRow extends TableRow<ScheduleEvent> {
+
+	/**
+	 * 20 days in milliseconds.
+	 */
+	private static final double MILLIS_UNTIL_IMPORTANT = 1728e6;
+	private static final Color EMPTY_CELL_COLOR = Color.TRANSPARENT, DEFAULT_START_COLOR = new Color(0, 0, 0, 0.3),
+			DEFAULT_END_COLOR = Color.GOLD, COMPLETE_COLOR = Color.GREEN, URGENT_END_COLOR = Color.HOTPINK,
+			PAST_DUE_COLOR = Color.RED, URGENT_PAST_DUE_COLOR = new Color(0.294117647, 0, 0.50980392156, 1);
+
+	private static final Background buildBackground(Color color) {
+		return new Background(new BackgroundFill(color, null, null));
+	}
+
+	private static final Color getColorFromDueDate(long time, Color upcomingColor, Color nearDueColor, Color dueColor) {
+
+		// Get the time until the date is due.
+		long timeUntilDue = time - System.currentTimeMillis();
+
+		// If the due date has passed, make the event red.
+		if (timeUntilDue < 0)
+			return dueColor;
+
+		// Interpolate the color between dueColor and our background color; the
+		// color will approach dueColor as the timeUntilDue approaches 0.
+		return nearDueColor.interpolate(upcomingColor, timeUntilDue / MILLIS_UNTIL_IMPORTANT);
+
+	}
+
+	private static final Color getColorFromDueDate(long time) {
+		return getColorFromDueDate(time, DEFAULT_START_COLOR, DEFAULT_END_COLOR.interpolate(Color.ORANGE, 0.6));
+	}
+
+	private static final Color getColorFromDueDate(long time, Color upcomingColor, Color nearDueColor) {
+		return getColorFromDueDate(time, upcomingColor, nearDueColor, PAST_DUE_COLOR);
+	}
+
+	private ScheduleModule module;
+
+	{
+
+		// Set the default background (so it isn't white)
+		setBackground(buildBackground(EMPTY_CELL_COLOR));
+
+		Label delete = new Label("Delete");
+		PopupWrapper<VBox> wrapper = PopupHelper.buildPopup(delete);
+		PopupHelper.applyRightClickPopup(getThis(), wrapper.popup);
+		delete.setOnMouseClicked(event -> {
+
+			if (!isEmpty() && getItem() != null && event.getButton() == MouseButton.PRIMARY) {
+				getItem().delete();
+				module.removeEvent(getItem());
+				wrapper.popup.hide();
+			}
+		});
+
+		setTextFill(Color.WHITE);
+
+		setOnMouseClicked(event -> {
+			if (event.getButton() == (MouseButton.PRIMARY) && !isEmpty()
+					&& !(event.getPickResult().getIntersectedNode() instanceof SelectableCell)) {
+				try {
+					WindowManager.setScene(new NewEvent(module, getItem()));
+					event.consume();
+				} catch (IOException | NotSwitchableException e) {
+					e.printStackTrace();
+				}
+			}
+		});
+
+		setOnMouseEntered(event -> {
+			if (!isEmpty()) {
+				setBackground(buildBackground(Color.WHITE));
+				setTextFill(Color.BLACK);
+			}
+		});
+
+		setOnMouseExited(event -> {
+			resetBackground();
+		});
+	}
+
+	public ScheduleRow(ScheduleModule module) {
+		this.module = module;
+	}
+
+	private TableRow<ScheduleEvent> getThis() {
+		return this;
+	}
+
+	private void resetBackground() {
+		if (isEmpty() || getItem() == null) {
+			setBackground(buildBackground(Color.TRANSPARENT));
+			return;
+		}
+
+		setBackground(
+				buildBackground(getItem().complete.get() ? COMPLETE_COLOR
+						: (getItem().urgent.get() ? getColorFromDueDate(getItem().dueDate.get(), DEFAULT_START_COLOR,
+								URGENT_END_COLOR, URGENT_PAST_DUE_COLOR)
+								: getColorFromDueDate(getItem().dueDate.get()))));
+		setTextFill(Color.WHITE);
+	}
+
+	@Override
+	protected void updateItem(ScheduleEvent item, boolean empty) {
+		if (item == getItem())
+			return;
+		super.updateItem(item, empty);
+
+		if (item == null || empty) {
+			setBackground(buildBackground(EMPTY_CELL_COLOR));
+		} else {
+			resetBackground();
+		}
+	}
+
+}
