+++ src/guis/zeale/guis/schedule_module/ScheduleModule.java
@@ -6,27 +6,18 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 
 import javafx.beans.property.BooleanProperty;
-import javafx.beans.value.ObservableValue;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
-import javafx.scene.control.Label;
 import javafx.scene.control.TableCell;
 import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableRow;
 import javafx.scene.control.TableView;
-import javafx.scene.control.cell.CheckBoxTableCell;
-import javafx.scene.input.MouseButton;
-import javafx.scene.layout.Background;
-import javafx.scene.layout.BackgroundFill;
 import javafx.scene.layout.Pane;
-import javafx.scene.layout.VBox;
 import javafx.scene.paint.Color;
 import javafx.util.Callback;
 import krow.guis.GUIHelper;
-import krow.guis.PopupHelper;
-import krow.guis.PopupHelper.PopupWrapper;
 import krow.guis.schedule_module.ScheduleEvent;
+import krow.guis.schedule_module.ScheduleRow;
 import krow.guis.schedule_module.SelectableCell;
 import kröw.core.Kröw;
 import kröw.core.managers.WindowManager;
@@ -39,14 +30,8 @@ public class ScheduleModule extends Page {
 	private static final ObservableList<ScheduleEvent> events = FXCollections.observableArrayList();
 
 	static {
-		// TODO Delete this.
-		System.setErr(Kröw.deferr);
-		System.setOut(Kröw.defout);
 		importData();
 	}
-
-	private static final Color EMPTY_CELL_COLOR = Color.TRANSPARENT, DEFAULT_START_COLOR = new Color(0, 0, 0, 0.3),
-			DEFAULT_END_COLOR = Color.GOLD, COMPLETE_COLOR = Color.GREEN, URGENT_END_COLOR = Color.HOTPINK;
 
 	public ObservableList<ScheduleEvent> getEventList() {
 		return events;
@@ -82,39 +67,7 @@ public class ScheduleModule extends Page {
 		events.sort(null);
 	}
 
-	private static final void overwriteImport() {
-		events.clear();
-		importData();
-	}
-
-	/**
-	 * 20 days in milliseconds.
-	 */
-	private static final double MILLIS_UNTIL_IMPORTANT = 1728e6;
-
-	private static final Background buildBackground(Color color) {
-		return new Background(new BackgroundFill(color, null, null));
-	}
-
 	private static final SimpleDateFormat dateFormatter = new SimpleDateFormat("MM/dd/yy"/* + " ~ hh:mm:ss" */);
-
-	private static final Color getColorFromDueDate(long time) {
-		return getColorFromDueDate(time, DEFAULT_START_COLOR, DEFAULT_END_COLOR.interpolate(Color.ORANGE, 0.6));
-	}
-
-	private static final Color getColorFromDueDate(long time, Color upcomingColor, Color dueColor) {
-
-		// Get the time until the date is due.
-		long timeUntilDue = time - System.currentTimeMillis();
-
-		// If the due date has passed, make the event red.
-		if (timeUntilDue < 0)
-			return Color.RED;
-
-		// Interpolate the color between dueColor and our background color; the
-		// color will approach dueColor as the timeUntilDue approaches 0.
-		return dueColor.interpolate(upcomingColor, timeUntilDue / MILLIS_UNTIL_IMPORTANT);
-	}
 
 	@FXML
 	private Pane root;
@@ -153,86 +106,7 @@ public class ScheduleModule extends Page {
 		urgencyColumn.setCellValueFactory(param -> param.getValue().urgent);
 		completeColumn.setCellValueFactory(param -> param.getValue().complete);
 
-		eventTable.setRowFactory(new Callback<TableView<ScheduleEvent>, TableRow<ScheduleEvent>>() {
-			@Override
-			public TableRow<ScheduleEvent> call(TableView<ScheduleEvent> param) {
-				return new TableRow<ScheduleEvent>() {
-
-					private void resetBackground() {
-						if (isEmpty() || getItem() == null) {
-							setBackground(buildBackground(Color.TRANSPARENT));
-							return;
-						}
-
-						setBackground(buildBackground(getItem().complete.get() ? COMPLETE_COLOR
-								: (getItem().urgent.get() ? getColorFromDueDate(getItem().dueDate.get(),
-										DEFAULT_START_COLOR, URGENT_END_COLOR)
-										: getColorFromDueDate(getItem().dueDate.get()))));
-						setTextFill(Color.WHITE);
-					}
-
-					private TableRow<ScheduleEvent> getThis() {
-						return this;
-					}
-
-					{
-
-						// Set the default background (so it isn't white)
-						setBackground(buildBackground(EMPTY_CELL_COLOR));
-
-						Label delete = new Label("Delete");
-						PopupWrapper<VBox> wrapper = PopupHelper.buildPopup(delete);
-						PopupHelper.applyRightClickPopup(getThis(), wrapper.popup);
-						delete.setOnMouseClicked(event -> {
-
-							if (!isEmpty() && getItem() != null && event.getButton() == MouseButton.PRIMARY) {
-								getItem().delete();
-								events.remove(getItem());
-								wrapper.popup.hide();
-							}
-						});
-
-						setTextFill(Color.WHITE);
-
-						setOnMouseClicked(event -> {
-							if (event.getButton() == (MouseButton.PRIMARY) && !isEmpty()
-									&& !(event.getPickResult().getIntersectedNode() instanceof SelectableCell)) {
-								try {
-									WindowManager.setScene(new NewEvent(ScheduleModule.this, getItem()));
-									event.consume();
-								} catch (IOException | NotSwitchableException e) {
-									e.printStackTrace();
-								}
-							}
-						});
-
-						setOnMouseEntered(event -> {
-							if (!isEmpty()) {
-								setBackground(buildBackground(Color.WHITE));
-								setTextFill(Color.BLACK);
-							}
-						});
-
-						setOnMouseExited(event -> {
-							resetBackground();
-						});
-					}
-
-					@Override
-					protected void updateItem(ScheduleEvent item, boolean empty) {
-						if (item == getItem())
-							return;
-						super.updateItem(item, empty);
-
-						if (item == null || empty) {
-							setBackground(buildBackground(EMPTY_CELL_COLOR));
-						} else {
-							resetBackground();
-						}
-					}
-				};
-			}
-		});
+		eventTable.setRowFactory(param -> new ScheduleRow(ScheduleModule.this));
 
 		dateColumn.setCellFactory(param -> {
 			return new TableCell<ScheduleEvent, Number>() {
@@ -244,10 +118,6 @@ public class ScheduleModule extends Page {
 
 				public void bindTextFill() {
 					getTableRow().textFillProperty().bindBidirectional(textFillProperty());
-				}
-
-				public void unbindTextFill() {
-					getTableRow().textFillProperty().unbindBidirectional(textFillProperty());
 				}
 
 				protected void updateItem(Number item, boolean empty) {
@@ -277,10 +147,6 @@ public class ScheduleModule extends Page {
 
 				public void bindTextFill() {
 					getTableRow().textFillProperty().bindBidirectional(textFillProperty());
-				}
-
-				public void unbindTextFill() {
-					getTableRow().textFillProperty().unbindBidirectional(textFillProperty());
 				}
 
 				protected void updateItem(String item, boolean empty) {
